Refactor: Relocate `DistributionOrderCollector` and `DDOrderReferenceCollector` from `job.service` to `launchers` package. Update imports across related classes.
