+++ backend/de.metas.distribution.rest-api/src/main/java/de/metas/distribution/mobileui/launchers/DDOrderReferenceCollector.java
@@ -1,8 +1,10 @@
-package de.metas.distribution.mobileui.job.service;
+package de.metas.distribution.mobileui.launchers;
 
 import de.metas.distribution.ddorder.DDOrderId;
 import de.metas.distribution.mobileui.job.model.DDOrderReference;
 import de.metas.distribution.mobileui.job.model.DistributionJob;
+import de.metas.distribution.mobileui.job.service.DistributionJobLoader;
+import de.metas.distribution.mobileui.job.service.DistributionJobLoaderSupportingServices;
 import lombok.Builder;
 import lombok.NonNull;
 import org.eevolution.model.I_DD_Order;
